+++ packages/core/example/change-account-role.ts
@@ -1,7 +1,6 @@
 import {SDK} from '../src';
-import {MasterWallet} from '../src/wallet';
 import 'dotenv/config'
-import {BlockchainType} from '../src/blockchain';
+import { Role, Account } from "../src/accounts";
 
 async function main() {
   const sdk = new SDK({
@@ -10,13 +9,8 @@ async function main() {
     url: "http://localhost:8080/api/v1"
   });
 
-  const wallet: MasterWallet = await sdk.wallets.createMasterWallet(
-    "wallet1",
-    BlockchainType.Klaytn,
-    "passphrase"
-  );
-
-  console.log(wallet.getData());
+  const account: Account = await sdk.organizations.changeAccountRole("9efbb37b3cc2e78c389ad389ce9d05ef", Role.ADMIN);
+  console.log(account);
 }
 
 main().catch((e) => console.error(e));