+++ packages/core/example/change-user-wallet-name.ts
@@ -1,7 +1,6 @@
 import {SDK} from '../src';
 import {MasterWallet} from '../src/wallet';
 import 'dotenv/config'
-import {BlockchainType} from '../src/blockchain';
 
 async function main() {
   const sdk = new SDK({
@@ -10,12 +9,8 @@ async function main() {
     url: "http://localhost:8080/api/v1"
   });
 
-  const wallet: MasterWallet = await sdk.wallets.createMasterWallet(
-    "wallet1",
-    BlockchainType.Klaytn,
-    "passphrase"
-  );
-
+  const wallet: MasterWallet = await sdk.wallets.getMasterWallet("9efbb37b3cc2e78c389ad389ce9d05ef");
+  await wallet.changeName("name-changed");
   console.log(wallet.getData());
 }