+++ packages/core/example/get-access-token.ts
@@ -1,6 +1,6 @@
-import {SDK} from '../src';
+import { SDK } from "../src";
 import 'dotenv/config'
-import { BlockchainType } from "../src/blockchain";
+import { Token } from "../src/types";
 
 async function main() {
   const sdk = new SDK({
@@ -9,7 +9,8 @@ async function main() {
     url: "http://localhost:8080/api/v1"
   });
 
-  console.log(await sdk.organizations.getOrganizationBalance(BlockchainType.Klaytn));
+  const token: Token = await sdk.accounts.getAccessToken();
+  console.log(token);
 }
 
 main().catch((e) => console.error(e));