+++ packages/core/src/coins/index.ts
@@ -1,9 +1,13 @@
 export { Eth } from './eth';
 export { Klay } from './klay';
 export { Hib } from './hib';
+export { Hcut } from './hcut';
+export { Drm } from './drm';
 
-export enum Coins {
+export enum Ticker {
   Eth = 'Eth',
   Klay = 'Klay',
   Hib = 'Hib',
+  Hcut = 'Hcut',
+  Drm = 'Drm'
 }