+++ packages/core/src/organizations.ts
@@ -4,11 +4,11 @@ import {
   Balance, Secret, Token, Key,
 } from './types';
 
-import { Account } from './accounts';
-import { Blockchain } from './blockchain';
+import { Account, Role } from './accounts';
+import { BlockchainType } from './blockchain';
+import { BNConverter } from './utils';
 
 export interface Organization {
-  accessToken: string;
   henesisEthKey: Key;
   henesisKlayKey: Key;
   id: string;
@@ -21,15 +21,13 @@ export class Organizations {
 
   private readonly baseUrl = '/organizations';
 
-  private readonly DEFAULT_TOKEN_EXPIRED_TIME = 3600;
-
   constructor(client: Client) {
     this.client = client;
   }
 
-  public async getOrganizationBalance(blockchain: Blockchain): Promise<Balance> {
+  public async getOrganizationBalance(blockchain: BlockchainType): Promise<Balance> {
     const balance = await this.client.get(`${this.baseUrl}/balance?blockchain=${blockchain}`);
-    balance.amount = new BN(`${balance.amount}`);
+    balance.amount = BNConverter.hexStringToBN(balance.amount);
     return balance;
   }
 
@@ -45,9 +43,9 @@ export class Organizations {
     return this.client.post<Secret>(`${this.baseUrl}/secret`);
   }
 
-  public async createAccessToken(expiresIn?: number): Promise<Token> {
-    const requestExpiresIn = expiresIn || this.DEFAULT_TOKEN_EXPIRED_TIME;
-    return this.client.post<Token>(`${this.baseUrl}/token`, { expiresIn: requestExpiresIn });
+  public async changeAccountRole(accountId: string, role: Role): Promise<Account> {
+    return this.client.patch<Account>(`${this.baseUrl}/accounts/${accountId}`, {
+      role,
+    });
   }
-  // todo public async changeAccountRole()
 }