+++ packages/core/src/wallet.ts
@@ -2,16 +2,18 @@ import { Contract } from 'web3-eth-contract';
 import Web3 from 'web3';
 import { AbiItem } from 'web3-utils';
 import BN from 'bn.js';
+import CryptoJS from 'crypto-js';
 import { Client } from './sdk';
 import {
   PaginationOptions, Pagination, Key, KeyWithPriv, Balance,
 } from './types';
-import { Coin, MultiSigPayload } from './coin';
+import { Coin } from './coin';
 import { Keychains } from './keychains';
-import { Blockchain } from './blockchain';
+import { BlockchainType } from './blockchain';
 import { Factory, GlobalCoinFactoryGenerator } from './factory';
 import wallet from './contracts/MasterWallet.json';
-import { Converter } from './utils';
+import { BNConverter, ObjectConverter } from './utils';
+import { Ticker } from './coins';
 
 const Bytes = require('./vendor/eth-lib/bytes');
 const { keccak256 } = require('./vendor/eth-lib/hash');
@@ -25,9 +27,17 @@ export enum WalletStatus {
   Active = 'ACTIVE'
 }
 
+export interface MultiSigPayload {
+  walletAddress: string;
+  toAddress: string;
+  value: BN;
+  walletNonce: BN;
+  hexData: string;
+}
+
 export interface Transaction {
   id: string;
-  blockchain: Blockchain;
+  blockchain: BlockchainType;
   walletId: string;
   accountId: string;
   hash: string;
@@ -38,7 +48,7 @@ export interface WalletData {
   id: string;
   name: string;
   address: string;
-  blockchain: Blockchain;
+  blockchain: BlockchainType;
   createdAt: string;
   status: WalletStatus;
 }
@@ -46,13 +56,14 @@ export interface WalletData {
 export interface MasterWalletData extends WalletData {
   backupKey: Key;
   accountKey: Key;
+  encryptionKey: string;
 }
 
 export interface UserWalletData extends WalletData {
 
 }
 
-export interface UserWalletPaginationOptions extends PaginationOptions{
+export interface UserWalletPaginationOptions extends PaginationOptions {
   name?: string;
   id?: string;
   address?: string;
@@ -61,8 +72,8 @@ export interface UserWalletPaginationOptions extends PaginationOptions{
 function convertMultiSigPayloadToDTO(multiSigPayload: MultiSigPayload) {
   return {
     hexData: multiSigPayload.hexData,
-    walletNonce: multiSigPayload.walletNonce.toString(10),
-    value: multiSigPayload.value.toString(10),
+    walletNonce: BNConverter.bnToHexString(multiSigPayload.walletNonce),
+    value: BNConverter.bnToHexString(multiSigPayload.value),
     toAddress: multiSigPayload.toAddress,
     walletAddress: multiSigPayload.walletAddress,
   };
@@ -87,14 +98,14 @@ export abstract class Wallet {
     this.coinFactory = coinFactory;
   }
 
-  abstract getChain(): Blockchain;
+  abstract getChain(): BlockchainType;
 
   abstract verifyAddress(address: string): boolean;
 
   abstract isValidAddress(address: string): boolean;
 
   abstract transfer(
-    ticker: string,
+    ticker: Ticker | string,
     to: string,
     amount: BN,
     passphrase: string,
@@ -112,6 +123,10 @@ export abstract class Wallet {
   abstract getBalance(): Promise<Balance[]>;
 
   abstract getAddress(): string;
+
+  abstract getId(): string;
+
+  abstract changeName(name: string);
 }
 
 export abstract class EthLikeWallet extends Wallet {
@@ -130,7 +145,7 @@ export abstract class EthLikeWallet extends Wallet {
     return false;
   }
 
-  getChain(): Blockchain {
+  getChain(): BlockchainType {
     return this.masterWalletData.blockchain;
   }
 
@@ -162,7 +177,7 @@ export abstract class EthLikeWallet extends Wallet {
       signature,
       this.getChain(),
       multiSigPayload,
-      this.masterWalletData.id,
+      this.getId(),
       otpCode,
     );
   }
@@ -208,7 +223,7 @@ export abstract class EthLikeWallet extends Wallet {
   }
 
   async transfer(
-    ticker: string,
+    ticker: Ticker | string,
     to: string,
     amount: BN,
     passphrase: string,
@@ -220,7 +235,7 @@ export abstract class EthLikeWallet extends Wallet {
     const multiSigPayload: MultiSigPayload = {
       hexData,
       walletNonce: nonce,
-      value: new BN(0),
+      value: BNConverter.hexStringToBN('0x0'),
       toAddress: this.getAddress(),
       walletAddress: this.getAddress(),
     };
@@ -234,15 +249,17 @@ export abstract class EthLikeWallet extends Wallet {
       signature,
       this.getChain(),
       multiSigPayload,
-      this.masterWalletData.id,
+      this.getId(),
       otpCode,
     );
   }
 
   async getNonce(): Promise<BN> {
-    const nonce: Nonce = await this.client
-      .get<Nonce>(`${this.baseUrl}/${this.masterWalletData.id}/nonce`);
-    return new BN(`${nonce.nonce}`);
+    const nonce: {
+      nonce: string
+    } = await this.client
+      .get(`${this.baseUrl}/${this.masterWalletData.id}/nonce`);
+    return BNConverter.hexStringToBN(nonce.nonce);
   }
 }
 
@@ -256,6 +273,19 @@ export class MasterWallet extends EthLikeWallet {
   ) {
     super(client, walletData, keychains);
     this.wallet = new new Web3().eth.Contract((wallet as AbiItem[]));
+  }
+
+  async restorePassphrase(encryptedPassphrase: string, newPassphrase: string, otpCode?: string): Promise<void> {
+    const { encryptionKey } = this.masterWalletData;
+    const decrypted = CryptoJS.AES.decrypt(encryptedPassphrase, encryptionKey);
+    const passphrase = this.hex2a(decrypted.toString());
+    await this.changePassphrase(passphrase, newPassphrase, otpCode);
+  }
+
+  hex2a(hex: string): string {
+    let str = '';
+    for (let i = 0; i < hex.length; i += 2) str += String.fromCharCode(parseInt(hex.substr(i, 2), 16));
+    return str;
   }
 
   async changePassphrase(passphrase: string, newPassphrase: string, otpCode?: string): Promise<void> {
@@ -284,7 +314,7 @@ export class MasterWallet extends EthLikeWallet {
     const multiSigPayload: MultiSigPayload = {
       hexData: data,
       walletNonce: nonce,
-      value: new BN(0),
+      value: BNConverter.hexStringToBN('0x0'),
       toAddress: this.getAddress(),
       walletAddress: this.getAddress(),
     };
@@ -299,7 +329,7 @@ export class MasterWallet extends EthLikeWallet {
         `${this.baseUrl}/${this.masterWalletData.id}/user-wallets`,
         {
           name,
-          salt: salt.toString(10),
+          salt: BNConverter.bnToHexString(salt),
           signature,
           blockchain: this.getChain(),
           multiSigPayload: convertMultiSigPayloadToDTO(multiSigPayload),
@@ -328,7 +358,7 @@ export class MasterWallet extends EthLikeWallet {
   async getBalance(): Promise<Balance[]> {
     const balances: {
       coinType: string;
-      amount: number;
+      amount: string;
       name: string;
       symbol: string;
     }[] = await this.client
@@ -336,7 +366,7 @@ export class MasterWallet extends EthLikeWallet {
 
     return balances.map((balance) => ({
       symbol: balance.symbol,
-      amount: new BN(`${balance.amount}`),
+      amount: BNConverter.hexStringToBN(balance.amount),
       coinType: balance.coinType,
       name: balance.name,
     }));
@@ -351,8 +381,9 @@ export class MasterWallet extends EthLikeWallet {
   }
 
   async getUserWallets(options?: UserWalletPaginationOptions): Promise<Pagination<UserWallet>> {
-    const queryString: string = options ? Object.keys(Converter.toSnakeCase(options))
-      .map((key) => `${key}=${Converter.toSnakeCase(options)[key]}`).join('&') : '';
+    const queryString: string = options ? Object.keys(ObjectConverter.toSnakeCase(options))
+      .filter((key) => !!options[key])
+      .map((key) => `${key}=${ObjectConverter.toSnakeCase(options)[key]}`).join('&') : '';
 
     const data: Pagination<UserWalletData> = await this.client
       .get<Pagination<UserWalletData>>(`${this.baseUrl}/${this.masterWalletData.id}/user-wallets?${queryString}`);
@@ -366,6 +397,21 @@ export class MasterWallet extends EthLikeWallet {
         data,
       )),
     } as Pagination<UserWallet>;
+  }
+
+  getId(): string {
+    return this.masterWalletData.id;
+  }
+
+  async changeName(name: string) {
+    const masterWalletData: MasterWalletData = await this.client
+      .patch<MasterWalletData>(
+        `${this.baseUrl}/${this.masterWalletData.id}/name`,
+        {
+          name,
+        },
+      );
+    this.masterWalletData.name = masterWalletData.name;
   }
 }
 
@@ -383,15 +429,17 @@ export class UserWallet extends EthLikeWallet {
   }
 
   async getNonce(): Promise<BN> {
-    const nonce: Nonce = await this.client
-      .get<Nonce>(`${this.baseUrl}/${this.masterWalletData.id}/user-wallets/${this.userWalletData.id}/nonce`);
-    return new BN(`${nonce.nonce}`);
+    const nonce: {
+      nonce: string
+    } = await this.client
+      .get(`${this.baseUrl}/${this.masterWalletData.id}/user-wallets/${this.userWalletData.id}/nonce`);
+    return BNConverter.hexStringToBN(nonce.nonce);
   }
 
   async getBalance(): Promise<any[]> {
     const balances: {
       coinType: string;
-      amount: number;
+      amount: string;
       name: string;
       symbol: string;
     }[] = await this.client
@@ -399,7 +447,7 @@ export class UserWallet extends EthLikeWallet {
 
     return balances.map((balance) => ({
       symbol: balance.symbol,
-      amount: new BN(balance.amount.toLocaleString().replace(/,/g,'')),
+      amount: BNConverter.hexStringToBN(balance.amount),
       coinType: balance.coinType,
       name: balance.name,
     }));
@@ -411,5 +459,20 @@ export class UserWallet extends EthLikeWallet {
 
   getData(): UserWalletData {
     return this.userWalletData;
+  }
+
+  getId(): string {
+    return this.userWalletData.id;
+  }
+
+  async changeName(name: string) {
+    const userWalletData: UserWalletData = await this.client
+      .patch<UserWalletData>(
+        `${this.baseUrl}/${this.masterWalletData.id}/user-wallets/${this.userWalletData.id}/name`,
+        {
+          name,
+        },
+      );
+    this.userWalletData.name = userWalletData.name;
   }
 }