+++ packages/enclave/src/controllers/wallet.controller.ts
@@ -1,14 +1,12 @@
 import express from 'express';
-import {MasterWalletData, UserWallet, UserWalletData} from '@haechi-labs/henesis-wallet-core/lib/wallet';
-import {Pagination} from '@haechi-labs/henesis-wallet-core/lib/types';
-import {SDK} from '@haechi-labs/henesis-wallet-core';
-import BN from 'bn.js';
-import {Controller} from '../types';
+import {
+  MasterWalletData, Transaction, UserWallet, UserWalletData,
+} from '@haechi-labs/henesis-wallet-core/lib/wallet';
+import { SDK } from '@haechi-labs/henesis-wallet-core';
+import { BNConverter } from '@haechi-labs/henesis-wallet-core/lib/utils';
+import { Ticker } from '@haechi-labs/henesis-wallet-core/lib/coins';
+import { Controller } from '../types';
 import AbstractController from './controller';
-
-export interface Transaction {
-  transactionId: string;
-}
 
 export interface NonceResponse {
   nonce: string
@@ -35,8 +33,7 @@ export default class WalletController extends AbstractController implements Cont
     this.router.post(`${this.path}`, this.promiseWrapper(this.createMasterWallet));
 
     this.router.get(`${this.path}/:masterWalletId/user-wallets`,
-        this.promiseWrapper(this.getUserWallets)
-    );
+      this.promiseWrapper(this.getUserWallets));
 
     this.router.post(
         `${this.path}/:masterWalletId/user-wallets`,
@@ -44,13 +41,13 @@ export default class WalletController extends AbstractController implements Cont
     );
 
     this.router.post(
-        `${this.path}/:masterWalletId/contractCall`,
-        this.promiseWrapper(this.sendMasterWalletContractCall),
+      `${this.path}/:masterWalletId/contract-call`,
+      this.promiseWrapper(this.sendMasterWalletContractCall),
     );
 
     this.router.post(
-        `${this.path}/:masterWalletId/user-wallets/:userWalletId/contractCall`,
-        this.promiseWrapper(this.sendUserWalletContractCall),
+      `${this.path}/:masterWalletId/user-wallets/:userWalletId/contract-call`,
+      this.promiseWrapper(this.sendUserWalletContractCall),
     );
 
     this.router.get(
@@ -72,7 +69,6 @@ export default class WalletController extends AbstractController implements Cont
         `${this.path}/:masterWalletId/user-wallets/:userWalletId/nonce`,
         this.promiseWrapper(this.getUserWalletNonce),
     );
-
     this.router.get(
         `${this.path}/:masterWalletId/balance`,
         this.promiseWrapper(this.getMasterWalletBalance),
@@ -106,27 +102,27 @@ export default class WalletController extends AbstractController implements Cont
 
   private async getMasterWallets(req: express.Request): Promise<MasterWalletData[]> {
     const wallets = await req.sdk.wallets.getMasterWallets();
-    return wallets.map(x => x.getData());
+    return wallets.map((x) => x.getData());
   }
 
   private async createMasterWallet(req: express.Request): Promise<MasterWalletData> {
     return (await req.sdk
-        .wallets
-        .createMasterWallet(
-            req.body.name,
-            req.body.blockchain,
-            req.body.passphrase,
-        )).getData();
+      .wallets
+      .createMasterWallet(
+        req.body.name,
+        req.body.blockchain,
+        req.body.passphrase,
+      )).getData();
   }
 
   private async changePassphrase(req: express.Request): Promise<void> {
     const masterWallet = await req.sdk
-        .wallets
-        .getMasterWallet(req.params.masterWalletId);
+      .wallets
+      .getMasterWallet(req.params.masterWalletId);
     return (await masterWallet.changePassphrase(
-        req.body.passphrase,
-        req.body.newPassphrase,
-        req.body.otp,
+      req.body.passphrase,
+      req.body.newPassphrase,
+      req.body.otp,
     ));
   }
 
@@ -136,17 +132,16 @@ export default class WalletController extends AbstractController implements Cont
         .getMasterWallet(req.params.masterWalletId);
 
     return (await masterWallet.createUserWallet(
-        req.body.name,
-        req.body.passphrase,
-        req.body.salt,
+      req.body.name,
+      req.body.passphrase,
+      req.body.salt ? BNConverter.hexStringToBN(req.body.salt) : undefined,
     )).getData();
   }
 
   private async getMasterWallet(req: express.Request): Promise<MasterWalletData> {
     const masterWallet = await req.sdk
-        .wallets
-        .getMasterWallet(req.params.masterWalletId);
-
+      .wallets
+      .getMasterWallet(req.params.masterWalletId);
     return (masterWallet.getData());
   }
 
@@ -156,38 +151,33 @@ export default class WalletController extends AbstractController implements Cont
         .getMasterWallet(req.params.masterWalletId);
 
     const balances = await masterWallet.getBalance();
-    return balances.map(x => {
-      return {
-        coinType: x.coinType,
-        amount: x.amount.toString(),
-        name: x.name,
-        symbol: x.symbol
-      };
-    });
+
+    return balances.map((x) => ({
+      coinType: x.coinType,
+      amount: BNConverter.bnToHexString(x.amount),
+      name: x.name,
+      symbol: x.symbol,
+    }));
   }
 
   private async getUserWalletBalance(req: express.Request): Promise<BalanceResponse[]> {
-    const userWallet = await this.getUserWallet(
-        req.sdk,
-        req.params.masterWalletId,
-        req.params.userWalletId,
+    const userWallet = await this.getUserWalletByContext(
+      req.sdk,
+      req.params.masterWalletId,
+      req.params.userWalletId,
     );
     const balances = await userWallet.getBalance();
-    return balances.map(x => {
-      return {
-        coinType: x.coinType,
-        amount: x.amount.toString(),
-        name: x.name,
-        symbol: x.symbol
-      };
-    });
+    return balances.map((x) => ({
+      coinType: x.coinType,
+      amount: BNConverter.bnToHexString(x.amount),
+      name: x.name,
+      symbol: x.symbol,
+    }));
   }
 
   private async getUserWalletTokenBalance(req: express.Request): Promise<any> {
     const userWallet = await this.getUserWallet(
-        req.sdk,
-        req.params.masterWalletId,
-        req.params.userWalletId,
+        req,
     );
     const balances = await userWallet.getBalance();
     const hibBalance  = balances.filter(x => x.name === "hib")[0]
@@ -205,18 +195,18 @@ export default class WalletController extends AbstractController implements Cont
         .wallets
         .getMasterWallet(req.params.masterWalletId);
     return {
-      nonce: (await masterWallet.getNonce()).toString()
+      nonce: BNConverter.bnToHexString(await masterWallet.getNonce()),
     };
   }
 
   private async getUserWalletNonce(req: express.Request): Promise<NonceResponse> {
-    const userWallet = await this.getUserWallet(
-        req.sdk,
-        req.params.masterWalletId,
-        req.params.userWalletId,
+    const userWallet = await this.getUserWalletByContext(
+      req.sdk,
+      req.params.masterWalletId,
+      req.params.userWalletId,
     );
     return {
-      nonce: (await userWallet.getNonce()).toString()
+      nonce: BNConverter.bnToHexString(await userWallet.getNonce()),
     };
   }
 
@@ -226,34 +216,28 @@ export default class WalletController extends AbstractController implements Cont
         .getMasterWallet(req.params.masterWalletId);
 
     const transaction = await masterWallet.contractCall(
-        req.body.contractAddress,
-        new BN(`${req.body.value}`),
-        req.body.data,
-        req.body.passphrase,
+      req.body.contractAddress,
+      BNConverter.hexStringToBN(req.body.value),
+      req.body.data,
+      req.body.passphrase,
     );
 
-    return {
-      transactionId: transaction.id,
-    };
+    return transaction;
   }
 
   private async sendUserWalletContractCall(req: express.Request): Promise<Transaction> {
-    const userWallet = await this.getUserWallet(
-        req.sdk,
-        req.params.masterWalletId,
-        req.params.userWalletId,
+    const userWallet = await this.getUserWalletByContext(
+      req.sdk,
+      req.params.masterWalletId,
+      req.params.userWalletId,
     );
 
-    const transaction = await userWallet.contractCall(
-        req.body.contractAddress,
-        new BN(`${req.body.value}`),
-        req.body.data,
-        req.body.passphrase,
+    return await userWallet.contractCall(
+      req.body.contractAddress,
+      BNConverter.hexStringToBN(req.body.value),
+      req.body.data,
+      req.body.passphrase,
     );
-
-    return {
-      transactionId: transaction.id,
-    };
   }
 
   private async sendMasterWalletCoin(req: express.Request): Promise<Transaction> {
@@ -261,57 +245,52 @@ export default class WalletController extends AbstractController implements Cont
         .wallets
         .getMasterWallet(req.params.masterWalletId);
 
-    const transaction = await masterWallet.transfer(
+    return await masterWallet.transfer(
       req.body.ticker,
       req.body.to,
-      new BN(`${req.body.amount}`),
+      BNConverter.hexStringToBN(req.body.amount),
       req.body.passphrase,
       req.body.otp,
     );
-
-    return {
-      transactionId: transaction.id,
-    };
   }
 
   private async sendUserWalletCoin(req: express.Request): Promise<Transaction> {
-    const userWallet = await this.getUserWallet(
-        req.sdk,
-        req.params.masterWalletId,
-        req.params.userWalletId,
+    const userWallet = await this.getUserWalletByContext(
+      req.sdk,
+      req.params.masterWalletId,
+      req.params.userWalletId,
     );
-
-    const transaction = await userWallet.transfer(
-        req.body.ticker,
-        req.body.to,
-        new BN(`${req.body.amount}`),
-        req.body.passphrase,
-        req.body.otp,
+    return await userWallet.transfer(
+      Ticker[req.body.ticker],
+      req.body.to,
+      BNConverter.hexStringToBN(req.body.amount),
+      req.body.passphrase,
+      req.body.otp,
     );
-
-    return {
-      transactionId: transaction.id,
-    };
   }
 
-  private async getUserWallet(sdk: SDK, masterWalletId: string, userWalletId: string): Promise<UserWallet> {
+  private async getUserWallet(req: express.Request): Promise<UserWallet> {
+    return (await this.getUserWalletByContext(
+      req.sdk,
+      req.params.masterWalletId,
+      req.params.userWalletId,
+    ));
+  }
+
+  private async getUserWalletByContext(sdk: SDK, masterWalletId: string, userWalletId: string): Promise<UserWallet> {
     return (await sdk.wallets.getMasterWallet(masterWalletId))
         .getUserWallet(userWalletId);
   }
 
   private async getUserWallets(req: express.Request): Promise<UserWalletData[]> {
     const masterWallet = await req.sdk.wallets.getMasterWallet(req.params.masterWalletId);
-    let wallets;
-    if ( req.params.page )
-      wallets = (await masterWallet.getUserWallets({
-        page: +req.params.page,
-        size: +req.params.size,
-        sort: req.params.sort,
-        name: req.params.name,
-        address: req.params.address
-      })).results;
-    else
-      wallets = (await masterWallet.getUserWallets()).results;
-    return wallets.map(x => x.getData());
+    const wallets = (await masterWallet.getUserWallets({
+      page: +req.query.page,
+      size: +req.query.size,
+      sort: req.query.sort as string,
+      name: req.query.name as string,
+      address: req.query.address as string,
+    })).results;
+    return wallets.map((x) => x.getData());
   }
 }