+++ packages/enclave/src/middlewares/converter.ts
@@ -1,7 +1,7 @@
 import express from 'express';
-import { Converter } from '@haechi-labs/henesis-wallet-core/lib/utils';
+import { ObjectConverter } from '@haechi-labs/henesis-wallet-core/lib/utils';
 
 export default (req: express.Request, resp: express.Response, next: express.NextFunction) => {
-  req.body = Converter.toCamelCase(req.body);
+  req.body = ObjectConverter.toCamelCase(req.body);
   next();
 };